Proyecto CNN/requirements.txt
python 3.11
tensorflow==2.15.0
numpy
opencv-python
pillow
matplotlib
scikit-learn
bing-image-downloader
kagglehub
